refactor(ci): Update Firebase authentication method in workflow
This commit updates the authentication method used for deploying to Firebase App Distribution in the `build.yml` workflow.

The `token` parameter, which used a Firebase CI token, has been replaced with `serviceCredentialsFileContent`. This new approach uses a service account key in JSON format, which is a more secure and recommended way to authenticate with Firebase services.