refactor(android): Decouple Android app into separate module
This commit restructures the project by extracting the Android-specific code from the `composeApp` module into a new, dedicated `androidApp` module. This is done to address a deprecation warning that doesn't allow for the `com.android.application` plugin to be applied in the same module as the Kotlin Multiplatform plugin (`org.jetbrains.kotlin.multiplatform`).